refac: rag prompt template
